refactor(website): extract SwitchOrAddVersion
